\name{cork}
\alias{cork}
\docType{data}
\title{Cork dataset}
\description{
The cork dataset (Rao, 1948) consists of weights of cork borings taken by the author (Rao, C.R.) from the north (N), east (E), south (S) and west (W) directions of the trunk for 28 trees in a block of plantations.
}
\usage{data(cork)}
\format{
  A data frame with 28 observations on the following 4 variables.
  \describe{
    \item{\code{N}}{North direction}
    \item{\code{E}}{East direction}
    \item{\code{S}}{South direction}
    \item{\code{W}}{West direction}
  }
}

\source{
  C.R. Rao. Tests of significance in multivariate analysis.  Biometrika, 35:58-79, 1948. 
}

\examples{
  data(cork)
}
\keyword{datasets}
