\name{glasses}
\alias{glasses}
\docType{data}
\title{Glasses dataset}
\description{
These are glasses.
}
\usage{data(glasses)}
\format{
  A data frame with 73 observations on the following 14 variables.
  \describe{
    \item{\code{ID}}{a numeric vector}
    \item{\code{CLA}}{a numeric vector}
    \item{\code{SIO2}}{a numeric vector}
    \item{\code{AL2O3}}{a numeric vector}
    \item{\code{NA2O}}{a numeric vector}
    \item{\code{K2O}}{a numeric vector}
    \item{\code{CAO}}{a numeric vector}
    \item{\code{MGO}}{a numeric vector}
    \item{\code{FE2O3}}{a numeric vector}
    \item{\code{TIO2}}{a numeric vector}
    \item{\code{P2O5}}{a numeric vector}
    \item{\code{CL}}{a numeric vector}
    \item{\code{SO3}}{a numeric vector}
    \item{\code{MNO}}{a numeric vector}
  }
}
%\source{
%  ~~ reference to a publication or URL from which the data were obtained ~~
%}
%\references{
%  ~~ possibly secondary sources and usages ~~
%}

\keyword{datasets}
