\name{plot.localdepth}
\title{A DD-plot for Local Depth versus Depth}
\usage{
\method{plot}{localdepth}(x, xlab="Depth", ylab="Local Depth", main="DD plot", ...)
}

\alias{plot.localdepth}

\arguments{
 \item{x}{an object of class \code{localdepth}}
 \item{xlab}{a title for the x axis: see \code{\link{title}}}
 \item{ylab}{a title for the y axis: see \code{\link{title}}}
 \item{main}{an overall title for the plot: see \code{\link{title}}}
 \item{...}{graphical parameters can be given as arguments.}
}

\description{

}

\value{
The function return the \code{x} object as \code{invisible}.
}

\author{Claudio Agostinelli and Mario Romanazzi}

\seealso{\code{\link{localdepth}}}

\examples{


}

\keyword{kk}
\keyword{kk}


